Updated file organization
