+++ .gitignore
@@ -2,5 +2,4 @@
 /sign_cnn/.idea/other.xml
 /sign_cnn/.idea/inspectionProfiles/profiles_settings.xml
 /sign_cnn/.idea/inspectionProfiles/Project_Default.xml
-/sign_cnn/.idea/sign_cnn.iml
 *.jpg
