+++ 3-ResNet-Happy_faces_classification.py
@@ -141,33 +141,34 @@ def ResNet50(input_shape, output_shape):
     return model
 
 
-clf = ResNet50((64, 64, 3), 6)
-clf.compile(optimizer=tf.keras.optimizers.Adam(),
-            loss=tf.keras.losses.CategoricalCrossentropy(),
-            metrics=['accuracy'])
-print(clf.summary())
+if __name__ == '__main__':
+    clf = ResNet50((64, 64, 3), 6)
+    clf.compile(optimizer=tf.keras.optimizers.Adam(),
+                loss=tf.keras.losses.CategoricalCrossentropy(),
+                metrics=['accuracy'])
+    print(clf.summary())
 
-# Data preparation and loading
-x_train, y_train, x_test, y_test, _ = load_dataset("./datasets/train_signs.h5",
-                                                   "./datasets/test_signs.h5")
-x_train, x_test = x_train / 255., x_test / 255.
-y_train = tf.one_hot(y_train.T.squeeze(), depth=6).numpy()
-y_test = tf.one_hot(y_test.T.squeeze(), depth=6).numpy()
+    # Data preparation and loading
+    x_train, y_train, x_test, y_test, _ = load_dataset("./datasets/train_signs.h5",
+                                                       "./datasets/test_signs.h5")
+    x_train, x_test = x_train / 255., x_test / 255.
+    y_train = tf.one_hot(y_train.T.squeeze(), depth=6).numpy()
+    y_test = tf.one_hot(y_test.T.squeeze(), depth=6).numpy()
 
-# Training model for 20 epochs
-r = clf.fit(x_train, y_train, epochs=20, batch_size=32, validation_data=(x_test, y_test))
+    # Training model for 20 epochs
+    r = clf.fit(x_train, y_train, epochs=20, batch_size=32, validation_data=(x_test, y_test))
 
-# Plot training and validation loss and accuracy plots
-plot_loss_accuracy_vs_epoch(r, './happy_ResNet/loss_acc_plot.png')
+    # Plot training and validation loss and accuracy plots
+    plot_loss_accuracy_vs_epoch(r, './3-ResNet_plots.png')
 
-# Test with arbitrary image (Uncomment the following lines)
-# img_path = './2.jpg'  # change this to your image path
-# img = tf.keras.utils.load_img(img_path, target_size=(64, 64))
-# plt.imshow(img)
-#
-# d = tf.keras.utils.img_to_array(img)
-# d = np.expand_dims(d, axis=0)
-# d = tf.keras.applications.resnet50.preprocess_input(d)
-# plt.title(f'Predicted: {np.argmax(clf.predict(d)[0])}')
-# plt.show()
-# print(clf.predict(d)[0])
+    # Test with arbitrary image (Uncomment the following lines)
+    # img_path = './2.jpg'  # change this to your image path
+    # img = tf.keras.utils.load_img(img_path, target_size=(64, 64))
+    # plt.imshow(img)
+    #
+    # d = tf.keras.utils.img_to_array(img)
+    # d = np.expand_dims(d, axis=0)
+    # d = tf.keras.applications.resnet50.preprocess_input(d)
+    # plt.title(f'Predicted: {np.argmax(clf.predict(d)[0])}')
+    # plt.show()
+    # print(clf.predict(d)[0])
