--- sign_cnn/.idea/.gitignore
@@ -1,9 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-
-/misc.xml
